refactor: extract capability::guard (require_tier1 / require_explicit_server_url)
Fourth step of the capability.rs module split, stacked on the tier.rs
extraction. Pure lift-and-shift: inference_server_required_message,
require_tier1, and require_explicit_server_url (plus tests) move into
capability/guard.rs unchanged. capability.rs re-exports all three,
since callers throughout the crate use them via capability::require_*.

Depends on tier.rs (already extracted, for the Tier parameter type);
lands fourth.
